tidy(repo): rename graph updater to api
