--- api/.env.example
@@ -1,6 +1,0 @@
-PORT=4000
-HOST='0.0.0.0'
-DB_NAME=dataServices
-DB_URL="http://database:8529"
-DB_USER=root
-DB_PASS=yourpassword
