+++ api/docs/img/arango-graph-mutation-example.png
