--- api/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "api",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "type": "module",
-  "scripts": {
-    "test": "NODE_OPTIONS=--experimental-vm-modules jest --coverage",
-    "lint": "eslint src",
-    "dbg": "node --experimental-vm-modules --inspect-brk node_modules/.bin/jest --runInBand --no-cache",
-    "start": "node --experimental-vm-modules index.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@graphql-tools/schema": "^10.0.0",
-    "arangojs": "^8.4.0",
-    "dotenv-safe": "^8.2.0",
-    "graphql": "^16.7.1",
-    "graphql-type-json": "^0.3.2",
-    "graphql-yoga": "^4.0.4",
-    "nats": "^2.15.1"
-  },
-  "devDependencies": {
-    "jest": "^29.7.0"
-  }
-}
