--- api/src/schema/compliance-check.js
@@ -1,9 +1,0 @@
-import gql from 'graphql-tag';
-
-export const typeDef = gql`
-type ComplianceCheck {
-    checkPasses: Boolean
-    metadata: JSON
-}
-`
-
