--- api/src/schema/remote-repository-check.js
@@ -1,9 +1,0 @@
-import gql from 'graphql-tag';
-
-export const typeDef = gql`
-type RemoteRepositoryCheck {
-    vulnerabilityAlertsEnabled: ComplianceCheck
-    automatedSecurityFixesEnabled: ComplianceCheck
-    branchProtectionEnabled: ComplianceCheck
-}
-`
