--- api/src/schema/url-scanning-check.js
@@ -1,14 +1,0 @@
-import gql from 'graphql-tag';
-
-export const typeDef = gql`
-type URLScanningCheck {
-    accessibility: AccessibilityReport
-}
-
-type AccessibilityReport {
-   objectAlt: ComplianceCheck 
-   areaAlt: ComplianceCheck
-   ariaAllowedAttribute: ComplianceCheck
-}
-`
-
